Refactor `SearchPath` to be `Bson` instead of `ToBsonValue`
JAVA-4415
