--- driver-core/src/main/com/mongodb/client/model/ToBsonValue.java
@@ -1,32 +1,0 @@
-/*
- * Copyright 2008-present MongoDB, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.mongodb.client.model;
-
-import com.mongodb.annotations.Evolving;
-import org.bson.BsonValue;
-
-/**
- * An interface for types that are able to render themselves into a {@link BsonValue}.
- */
-@Evolving
-public interface ToBsonValue {
-    /**
-     * Renders into {@link BsonValue}.
-     *
-     * @return A {@link BsonValue} representation.
-     */
-    BsonValue toBsonValue();
-}
